+++ app/__init__.py
@@ -1,0 +1,9 @@
+"""
+Application layer - Contains application services and factories
+"""
+
+from .broker_factory import create_broker
+
+__all__ = [
+    "create_broker",
+]